Fix incorrect nested Dialogs behaviour (#489)
* add tests to verify the nested Dialog behaviour

* set mounted to true once rendered once

* cache useWindowEvent listener

We only care about the very last version of the listener function. This
allows us to only change the event listener if the event name (string)
and options (boolean | object) change.

* add/delete messages when mounting/unmounting

We don't require a dedicated hook anymore, so this is a bit of cleanup!

* add comments to the FocusResult enum

* splitup functionality and make it a bit more clear using feature flags

* add getDialogOverlays helper

* simplify the Portal component

We don't need to add the current element to the Stack. We only want to
take care of that in the Dialog component itself.

* drop dom-containers

Currently it is only used in a single spot, so I inlined it into that
file.

* simplify the FocusTrap component, use new API

* improve Dialog component

* update CHANGELOG